UI math completed-- TODO
Created an Optimism pool to confirm calculation of how liquidity in a position is distributed based on its asset distribution. And have confirmed the math with Uniswap Research.

Next step is to spread a fixed amount of liquidity in a hypothetical pool, and simulate swaps within and across ticks. Then use the Optimism pool to confirm the math (this pool has a single position).

After that; aggregate up positions across ticks and simulate; then confirm in OP pool.

Once we can successfully simulate swaps and fees across different marginal positions, we can begin retroactive analysis & modeling :)